Class work 20151007
Base for the class work last 20151027: a tic tac toe game between the
user and the machine.
Currently it's working on the base.
The user have to select the spaces and press another time to get the
machine move.
When the game is over, there won't be possible to do more movements and
the result will be display in the console (this is something to improve
=> make the line of the winner).
Currently the minimax algorithm is being executed for every move, it has
to be changed to move across the initial tree according to the movements
of the user and its own.